Fixed bug in div assembly function. Handles negative numbers correctly. Fixed bug in printf function, negative numbers now print correctly.

+++ sw/mini-libc/printf.c
@@ -203,7 +203,7 @@ int prints(char** dst, const char *string, int width, int pad)
 
 
 /* Printf an integer */
-int printi(char** dst, unsigned long i, int b, int sg, int width, int pad, int letbase)
+int printi(char** dst, int i, int b, int sg, int width, int pad, int letbase)
 {
     char print_buf[PRINT_BUF_LEN];
     char *s;
